requirements.txt
# RemoteHive Backend API Dependencies
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
# MongoDB Atlas only - SQLAlchemy and Alembic removed for MongoDB-only architecture

# MongoDB dependencies
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver
beanie==1.24.0  # Async ODM for MongoDB
dnspython==2.4.2  # Required for MongoDB SRV connections

# Authentication and security
python-jose==3.3.0
cryptography>=3.4.8,<42.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Data validation and settings
pydantic==2.9.2
pydantic-settings==2.6.1

# Background tasks and caching
celery==5.3.4
redis==5.0.1
flower==2.0.1

# HTTP requests and utilities
requests==2.31.0
httpx==0.24.1

# Web scraping and browser automation (for backend integration)
playwright==1.40.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Configuration and environment
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.2

# Email templates and rendering
jinja2==3.1.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Additional production dependencies
gunicorn==21.2.0  # For production WSGI server
sentry-sdk[fastapi]==1.38.0  # Error tracking